don't allow the material farmer stats to affect the item printer's stats
