+++ SerialPrograms/Source/PokemonSV/Programs/Farming/PokemonSV_MaterialFarmer.h
@@ -22,33 +22,6 @@ namespace PokemonAutomation{
 namespace NintendoSwitch{
 namespace PokemonSV{
 
-class LetsGoEncounterBotTracker;
-
-
-class MaterialFarmer_Descriptor : public SingleSwitchProgramDescriptor{
-public:
-    MaterialFarmer_Descriptor();
-
-    struct Stats : public LetsGoEncounterBotStats{
-        Stats()
-            : m_sandwiches(m_stats["Sandwiches"])
-            , m_autoheals(m_stats["Auto Heals"])
-            , m_game_resets(m_stats["Game Resets"])
-            , m_errors(m_stats["Errors"])
-        {
-            m_display_order.insert(m_display_order.begin() + 2, {"Sandwiches", HIDDEN_IF_ZERO});
-            m_display_order.insert(m_display_order.begin() + 3, {"Auto Heals", HIDDEN_IF_ZERO});
-            m_display_order.insert(m_display_order.begin() + 4, {"Game Resets", HIDDEN_IF_ZERO});
-            m_display_order.insert(m_display_order.begin() + 5, {"Errors", HIDDEN_IF_ZERO});
-        }
-        std::atomic<uint64_t>& m_sandwiches;
-        std::atomic<uint64_t>& m_autoheals;
-        std::atomic<uint64_t>& m_game_resets;
-        std::atomic<uint64_t>& m_errors;
-    };
-    virtual std::unique_ptr<StatsTracker> make_stats() const override;
-};
-
 
 class MaterialFarmer : public SingleSwitchProgramInstance{
 public:
