+++ SerialPrograms/Source/PokemonSV/Programs/Farming/PokemonSV_MaterialFarmerTools.cpp
@@ -28,7 +28,6 @@
 #include "PokemonSV/Programs/Battles/PokemonSV_Battles.h"
 #include "PokemonSV/Programs/Sandwiches/PokemonSV_SandwichRoutines.h"
 #include "PokemonSV/Programs/ShinyHunting/PokemonSV_LetsGoTools.h"
-#include "PokemonSV_MaterialFarmer.h"
 #include "PokemonSV_MaterialFarmerTools.h"
 #include "PokemonSV/Programs/ShinyHunting/PokemonSV_ShinyHunt-Scatterbug.h"
 #include "NintendoSwitch/Programs/NintendoSwitch_SnapshotDumper.h"
@@ -192,11 +191,27 @@ MaterialFarmerOptions::MaterialFarmerOptions(
     }
 }
 
+MaterialFarmer_Descriptor::MaterialFarmer_Descriptor()
+    : SingleSwitchProgramDescriptor(
+        "PokemonSV:MaterialFarmer",
+        STRING_POKEMON + " SV", "Material Farmer",
+        "ComputerControl/blob/master/Wiki/Programs/PokemonSV/MaterialFarmer.md",
+        "Farm materials - Happiny dust from Chanseys/Blisseys, for Item Printer.",
+        FeedbackType::VIDEO_AUDIO,
+        AllowCommandsWhenRunning::DISABLE_COMMANDS,
+        PABotBaseLevel::PABOTBASE_12KB
+    )
+{}
 
-void run_material_farmer(SingleSwitchProgramEnvironment& env, BotBaseContext& context, MaterialFarmerOptions& options){
+std::unique_ptr<StatsTracker> MaterialFarmer_Descriptor::make_stats() const{
+    return std::unique_ptr<StatsTracker>(new Stats());
+}
+
+
+void run_material_farmer(SingleSwitchProgramEnvironment& env, BotBaseContext& context, 
+    MaterialFarmerOptions& options, MaterialFarmer_Descriptor::Stats& stats
+){
     
-    MaterialFarmer_Descriptor::Stats& stats = env.current_stats<MaterialFarmer_Descriptor::Stats>();
-
     LetsGoEncounterBotTracker encounter_tracker(
         env, env.console, context,
         stats,
